% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\docType{data}
\name{test}
\alias{test}
\title{A simulation data to test cutpointsOEHR}
\format{a dataframe contains 200 rows and 4 variables. The 4 varibles are
\describe{
 \item{t}{simulated times of developing survival outcomes like deathes, relapes, etc.}
 \item{d}{censoring indicator, 1 means that survival outcomes are not observed, 0 means survival outcomes are observed. The censoring proportion is set to be 20 percent. }
 \item{x}{a continuous variable which has U-shaped relationship with log relative hazard. }
 \item{x1}{a continuous variable which has linear relationship with log relative hazard. }
}}
\usage{
test
}
\description{
A dataframe named 'test'contains simulated (t,d,x,x1). The relationship of log relative hazard and x is set to to quandratic, which results in a U-shaped relationship.
}
\keyword{datasets}
